pyproject.toml
[project]
name = "voice-agent-streaming-demo"
version = "0.1.0"
description = "Event-driven streaming agent-assist demo (Redis Streams + FastAPI) using a light hexagonal architecture."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "redis>=5.0",
  "pyyaml>=6.0",
  "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "ruff>=0.3",
  "requests>=2.31",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "scikit-learn>=1.5",
  "joblib>=1.4",
  "pre-commit>=3.7"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
